\name{pasR}
\alias{pasR}
\alias{pasR}
\docType{package}
\title{
R Tools for Photoacoustic Spectroscopy
}
\description{
Basic fucntions for working with photoacoustic data.
}
\details{
\tabular{ll}{
Package: \tab pasR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-02-25\cr
License: \tab GPL-3\cr
}
}
\author{
Al Fischer <al@uga.edu>
}

\keyword{ package }